backend/requirements-backend.txt
## auth requirement
python-jose[cryptography]
passlib[argon2]
python-multipart
pydantic-settings
fastapi
sqlmodel
strawberry-graphql[fastapi]
uvicorn
psycopg2-binary

